+++ publish/CHANGELOG.md
@@ -1,6 +1,7 @@
 - v1.51
   - Adds new parameter `<cid>` to get the character id of the client that controls the object.
   - Adds new parameter `<owner>` to spawn data to override the initial owner assignment.
+  - Adds API for developers to register own parameter handlers.
   - Obsoletes parameter `<pchar>` as it returned wrong information anyways.
 
 - v1.50